grammar: proper atoms, vars, strings, and whitespace
* Implement proper rules for atoms, strings, and vars

This implements the syntax as recognised by erl_scan. It accepts
some ranges of latin1 characters as valid identifiers. Additionally
the same escape syntax is allowed in quoted atoms and strings.

Macro names can also be atoms, not only variables.

* Implement proper whitespace

This implements whitespace as recognised by erl_scan.
Additionally it leverages the "extras" feature to allow comments
to appear anywhere in the code.